extract into form526_workflow.rb
