+++ app/models/form526_submission.rb
@@ -4,6 +4,7 @@ require 'sentry_logging'
 
 class Form526Submission < ApplicationRecord
   include SentryLogging
+  include Form526Workflow
 
   # A 526 disability compensation form record. This class is used to persist the post transformation form
   # and track submission workflow steps.
@@ -37,14 +38,7 @@ class Form526Submission < ApplicationRecord
 
   validates(:auth_headers_json, presence: true)
 
-  FORM_526 = 'form526'
-  FORM_526_UPLOADS = 'form526_uploads'
-  FORM_4142 = 'form4142'
-  FORM_0781 = 'form0781'
-  FORM_8940 = 'form8940'
-  FLASHES = 'flashes'
   BIRLS_KEY = 'va_eauth_birlsfilenumber'
-  SUBMIT_FORM_526_JOB_CLASSES = %w[SubmitForm526AllClaim SubmitForm526].freeze
 
   def start
     if single_issue_hypertension_claim? && Flipper.enabled?(:disability_hypertension_compensation_fast_track)
@@ -65,30 +59,6 @@ class Form526Submission < ApplicationRecord
     Rails.logger.error 'The fast track was skipped due to the following error ' \
                        " and start_evss_submission wass called: #{e}"
     start_evss_submission(nil, { 'submission_id' => id })
-  end
-
-  # Kicks off a 526 submit workflow batch. The first step in a submission workflow is to submit
-  # an increase only or all claims form. Once the first job succeeds the batch will callback and run
-  # one (cleanup job) or more ancillary jobs such as uploading supporting evidence or submitting ancillary forms.
-  #
-  # @return [String] the job id of the first job in the batch, i.e the 526 submit job
-  #
-
-  def start_evss_submission(_status, options)
-    submission = Form526Submission.find(options['submission_id'])
-    id = submission.id
-    workflow_batch = Sidekiq::Batch.new
-    workflow_batch.on(
-      :success,
-      'Form526Submission#perform_ancillary_jobs_handler',
-      'submission_id' => id,
-      'first_name' => get_first_name
-    )
-    jids = workflow_batch.jobs do
-      EVSS::DisabilityCompensationForm::SubmitForm526AllClaim.perform_async(id)
-    end
-
-    jids.first
   end
 
   # Runs the start method above but first looks to see if the veteran has BIRLS IDs that previous start
@@ -238,147 +208,12 @@ class Form526Submission < ApplicationRecord
     @auth_headers_hash = nil # reset cache
   end
 
-  # The workflow batch success handler
-  #
-  # @param _status [Sidekiq::Batch::Status] the status of the batch
-  # @param options [Hash] payload set in the workflow batch
-  #
-  def perform_ancillary_jobs_handler(_status, options)
-    submission = Form526Submission.find(options['submission_id'])
-    # Only run ancillary jobs if submission succeeded
-    submission.perform_ancillary_jobs(options['first_name']) if submission.jobs_succeeded?
-  end
-
-  def jobs_succeeded?
-    a_submit_form_526_job_succeeded? && all_other_jobs_succeeded_if_any?
-  end
-
   class SubmitForm526JobStatusesError < StandardError; end
-
-  def a_submit_form_526_job_succeeded?
-    submit_form_526_job_statuses = form526_job_statuses.where(job_class: SUBMIT_FORM_526_JOB_CLASSES).order(:updated_at)
-    submit_form_526_job_statuses.presence&.any?(&:success?)
-  ensure
-    successful = submit_form_526_job_statuses.where(status: 'success').load
-    warn = ->(message) { log_message_to_sentry(message, :warn, { form_526_submission_id: id }) }
-    warn.call 'There are multiple successful SubmitForm526 job statuses' if successful.size > 1
-    if successful.size == 1 && submit_form_526_job_statuses.last.unsuccessful?
-      warn.call "There is a successful SubmitForm526 job, but it's not the most recent SubmitForm526 job"
-    end
-  end
-
-  def all_other_jobs_succeeded_if_any?
-    form526_job_statuses.where.not(job_class: SUBMIT_FORM_526_JOB_CLASSES).all?(&:success?)
-  end
-
-  # Creates a batch for the ancillary jobs, sets up the callback, and adds the jobs to the batch if necessary
-  #
-  # @param first_name [String] the first name of the user that submitted Form526
-  # @return [String] the workflow batch id
-  #
-  def perform_ancillary_jobs(first_name)
-    workflow_batch = Sidekiq::Batch.new
-    workflow_batch.on(
-      :success,
-      'Form526Submission#workflow_complete_handler',
-      'submission_id' => id,
-      'first_name' => first_name
-    )
-    workflow_batch.jobs do
-      submit_uploads if form[FORM_526_UPLOADS].present?
-      submit_form_4142 if form[FORM_4142].present?
-      submit_form_0781 if form[FORM_0781].present?
-      submit_form_8940 if form[FORM_8940].present?
-      upload_bdd_instructions if bdd?
-      submit_flashes if form[FLASHES].present?
-      cleanup
-    end
-  end
-
-  # Checks if all workflow steps were successful and if so marks it as complete.
-  #
-  # @param _status [Sidekiq::Batch::Status] the status of the batch
-  # @param options [Hash] payload set in the workflow batch
-  #
-  def workflow_complete_handler(_status, options)
-    submission = Form526Submission.find(options['submission_id'])
-    if submission.jobs_succeeded?
-      submission.send_form526_confirmation_email(options['first_name'])
-      submission.workflow_complete = true
-      submission.save
-    else
-      submission.send_form526_submission_failed_email(options['first_name'])
-    end
-  end
-
-  def send_form526_confirmation_email(first_name)
-    email_address = form['form526']['form526']['veteran']['emailAddress']
-    personalization_parameters = {
-      'email' => email_address,
-      'submitted_claim_id' => submitted_claim_id,
-      'date_submitted' => created_at.strftime('%B %-d, %Y %-l:%M %P %Z').sub(/([ap])m/, '\1.m.'),
-      'first_name' => first_name
-    }
-    Form526ConfirmationEmailJob.perform_async(personalization_parameters)
-  end
-
-  def send_form526_submission_failed_email(first_name)
-    email_address = form['form526']['form526']['veteran']['emailAddress']
-    personalization_parameters = {
-      'email' => email_address,
-      'submitted_claim_id' => submitted_claim_id,
-      'date_submitted' => created_at.strftime('%B %-d, %Y %-l:%M %P %Z').sub(/([ap])m/, '\1.m.'),
-      'first_name' => first_name
-    }
-    Form526SubmissionFailedEmailJob.perform_async(personalization_parameters)
-  end
-
-  def bdd?
-    form.dig('form526', 'form526', 'bddQualified') || false
-  end
 
   def single_issue_hypertension_claim?
     disabilities = form.dig('form526', 'form526', 'disabilities')
     disabilities.count == 1 &&
       disabilities.first['disabilityActionType'] == 'INCREASE' &&
       disabilities.first['diagnosticCode'] == 7101
-  end
-
-  private
-
-  def submit_uploads
-    # Put uploads on a one minute delay because of shared workload with EVSS
-    uploads = form[FORM_526_UPLOADS]
-    delay = 60.seconds
-    uploads.each do |upload|
-      EVSS::DisabilityCompensationForm::SubmitUploads.perform_in(delay, id, upload)
-      delay += 15.seconds
-    end
-  end
-
-  def upload_bdd_instructions
-    # send BDD instructions
-    EVSS::DisabilityCompensationForm::UploadBddInstructions.perform_in(60.seconds, id)
-  end
-
-  def submit_form_4142
-    CentralMail::SubmitForm4142Job.perform_async(id)
-  end
-
-  def submit_form_0781
-    EVSS::DisabilityCompensationForm::SubmitForm0781.perform_async(id)
-  end
-
-  def submit_form_8940
-    EVSS::DisabilityCompensationForm::SubmitForm8940.perform_async(id)
-  end
-
-  def submit_flashes
-    user = User.find(user_uuid)
-    BGS::FlashUpdater.perform_async(id) if user && Flipper.enabled?(:disability_compensation_flashes, user)
-  end
-
-  def cleanup
-    EVSS::DisabilityCompensationForm::SubmitForm526Cleanup.perform_async(id)
   end
 end
