+++ spec/models/form526_submission_spec.rb
@@ -440,10 +440,11 @@ RSpec.describe Form526Submission do
     end
   end
 
-  describe '#perform_ancillary_jobs_handler' do
+  describe 'BatchDoneHandlers#perform_ancillary_jobs_handler' do
     let(:status) { OpenStruct.new(parent_bid: SecureRandom.hex(8)) }
 
     context 'with an ancillary job' do
+      let(:workflow_steps) { Form526Workflow::BatchDoneHandlers.new }
       let(:form_json) do
         File.read('spec/support/disability_compensation_form/submissions/with_uploads.json')
       end
@@ -452,7 +453,7 @@ RSpec.describe Form526Submission do
         subject.form526_job_statuses <<
           Form526JobStatus.new(job_class: 'SubmitForm526AllClaim', status: 'success', job_id: 0)
         expect do
-          subject.perform_ancillary_jobs_handler(status, 'submission_id' => subject.id)
+          workflow_steps.perform_ancillary_jobs_handler(status, 'submission_id' => subject.id)
         end.to change(EVSS::DisabilityCompensationForm::SubmitUploads.jobs, :size).by(3)
       end
 
@@ -470,7 +471,7 @@ RSpec.describe Form526Submission do
           :warn,
           { form_526_submission_id: subject.id }
         )
-        subject.perform_ancillary_jobs_handler(status, 'submission_id' => subject.id)
+        workflow_steps.perform_ancillary_jobs_handler(status, 'submission_id' => subject.id)
       end
 
       it "warns when there's a successful submit526 job, but it's not the most recent submit526 job" do
@@ -488,7 +489,7 @@ RSpec.describe Form526Submission do
           :warn,
           { form_526_submission_id: subject.id }
         )
-        subject.perform_ancillary_jobs_handler(status, 'submission_id' => subject.id)
+        workflow_steps.perform_ancillary_jobs_handler(status, 'submission_id' => subject.id)
       end
     end
   end
@@ -660,7 +661,9 @@ RSpec.describe Form526Submission do
     end
   end
 
-  describe '#workflow_complete_handler' do
+  describe 'BatchDoneHandlers#workflow_complete_handler' do
+    let(:workflow_steps) { Form526Workflow::BatchDoneHandlers.new }
+
     describe 'success' do
       let(:options) do
         {
@@ -674,7 +677,7 @@ RSpec.describe Form526Submission do
 
         it 'sets the submission.complete to true' do
           expect(subject.workflow_complete).to be_falsey
-          subject.workflow_complete_handler(nil, 'submission_id' => subject.id)
+          workflow_steps.workflow_complete_handler(nil, 'submission_id' => subject.id)
           subject.reload
           expect(subject.workflow_complete).to be_truthy
         end
@@ -685,7 +688,7 @@ RSpec.describe Form526Submission do
 
         it 'sets the submission.complete to true' do
           expect(subject.workflow_complete).to be_falsey
-          subject.workflow_complete_handler(nil, 'submission_id' => subject.id)
+          workflow_steps.workflow_complete_handler(nil, 'submission_id' => subject.id)
           subject.reload
           expect(subject.workflow_complete).to be_truthy
         end
@@ -706,7 +709,7 @@ RSpec.describe Form526Submission do
             expect(args[0]['date_submitted']).to eql('July 20, 2012 2:15 p.m. UTC')
           end
 
-          subject.workflow_complete_handler(nil, options)
+          workflow_steps.workflow_complete_handler(nil, options)
         end
       end
 
@@ -725,7 +728,7 @@ RSpec.describe Form526Submission do
             expect(args[0]['date_submitted']).to eql('July 20, 2012 11:12 a.m. UTC')
           end
 
-          subject.workflow_complete_handler(nil, options)
+          workflow_steps.workflow_complete_handler(nil, options)
         end
       end
 
@@ -744,7 +747,7 @@ RSpec.describe Form526Submission do
             expect(args[0]['date_submitted']).to eql('July 20, 2012 8:07 a.m. UTC')
           end
 
-          subject.workflow_complete_handler(nil, options)
+          workflow_steps.workflow_complete_handler(nil, options)
         end
       end
 
@@ -753,7 +756,7 @@ RSpec.describe Form526Submission do
 
         it 'returns one job triggered' do
           expect do
-            subject.workflow_complete_handler(nil, 'submission_id' => subject.id)
+            workflow_steps.workflow_complete_handler(nil, 'submission_id' => subject.id)
           end.to change(Form526ConfirmationEmailJob.jobs, :size).by(1)
         end
       end
@@ -765,7 +768,7 @@ RSpec.describe Form526Submission do
 
         it 'sets the submission.complete to true' do
           expect(subject.workflow_complete).to be_falsey
-          subject.workflow_complete_handler(nil, 'submission_id' => subject.id)
+          workflow_steps.workflow_complete_handler(nil, 'submission_id' => subject.id)
           subject.reload
           expect(subject.workflow_complete).to be_falsey
         end
@@ -776,7 +779,7 @@ RSpec.describe Form526Submission do
 
         it 'sets the submission.complete to true' do
           expect(subject.workflow_complete).to be_falsey
-          subject.workflow_complete_handler(nil, 'submission_id' => subject.id)
+          workflow_steps.workflow_complete_handler(nil, 'submission_id' => subject.id)
           subject.reload
           expect(subject.workflow_complete).to be_falsey
         end
@@ -787,14 +790,14 @@ RSpec.describe Form526Submission do
 
         it 'returns zero jobs triggered' do
           expect do
-            subject.workflow_complete_handler(nil, 'submission_id' => subject.id)
+            workflow_steps.workflow_complete_handler(nil, 'submission_id' => subject.id)
           end.to change(Form526ConfirmationEmailJob.jobs, :size).by(0)
         end
       end
 
       it 'sends a submission failed email notification' do
         expect do
-          subject.workflow_complete_handler(nil, 'submission_id' => subject.id)
+          workflow_steps.workflow_complete_handler(nil, 'submission_id' => subject.id)
         end.to change(Form526SubmissionFailedEmailJob.jobs, :size).by(1)
       end
     end
